requirements.txt
Django~=5.0.6
djangorestframework~=3.15.1
PyJWT~=2.8.0
django-cors-headers~=    4.3.1